flash-dashboard/requirements.txt
streamlit==1.32.0
plotly==5.20.0
requests==2.31.0
pandas==2.2.1
pyyaml==6.0.1
